更改字段类型：Long 》 Integer

+++ ruoyi-modules/ruoyi-job/src/main/java/com/ruoyi/job/mapper/SysJobLogMapper.java
@@ -1,18 +1,18 @@
 package com.ruoyi.job.mapper;
 
 import java.util.List;
+
 import com.ruoyi.job.domain.SysJobLog;
 
 /**
  * 调度任务日志信息 数据层
- * 
+ *
  * @author ruoyi
  */
-public interface SysJobLogMapper
-{
+public interface SysJobLogMapper {
     /**
      * 获取quartz调度器日志的计划任务
-     * 
+     *
      * @param jobLog 调度日志信息
      * @return 调度任务日志集合
      */
@@ -27,15 +27,15 @@ public interface SysJobLogMapper
 
     /**
      * 通过调度任务日志ID查询调度信息
-     * 
+     *
      * @param jobLogId 调度任务日志ID
      * @return 调度任务日志对象信息
      */
-    public SysJobLog selectJobLogById(Long jobLogId);
+    public SysJobLog selectJobLogById(Integer jobLogId);
 
     /**
      * 新增任务日志
-     * 
+     *
      * @param jobLog 调度日志信息
      * @return 结果
      */
@@ -43,19 +43,19 @@ public interface SysJobLogMapper
 
     /**
      * 批量删除调度日志信息
-     * 
+     *
      * @param logIds 需要删除的数据ID
      * @return 结果
      */
-    public int deleteJobLogByIds(Long[] logIds);
+    public int deleteJobLogByIds(Integer[] logIds);
 
     /**
      * 删除任务日志
-     * 
+     *
      * @param jobId 调度日志ID
      * @return 结果
      */
-    public int deleteJobLogById(Long jobId);
+    public int deleteJobLogById(Integer jobId);
 
     /**
      * 清空任务日志
