+++ ruoyi-modules/ruoyi-system/src/main/java/com/ruoyi/system/mapper/SysConfigMapper.java
@@ -1,18 +1,18 @@
 package com.ruoyi.system.mapper;
 
 import java.util.List;
+
 import com.ruoyi.system.domain.SysConfig;
 
 /**
  * 参数配置 数据层
- * 
+ *
  * @author ruoyi
  */
-public interface SysConfigMapper
-{
+public interface SysConfigMapper {
     /**
      * 查询参数配置信息
-     * 
+     *
      * @param config 参数配置信息
      * @return 参数配置信息
      */
@@ -20,15 +20,15 @@ public interface SysConfigMapper
 
     /**
      * 通过ID查询配置
-     * 
+     *
      * @param configId 参数ID
      * @return 参数配置信息
      */
-    public SysConfig selectConfigById(Long configId);
+    public SysConfig selectConfigById(Integer configId);
 
     /**
      * 查询参数配置列表
-     * 
+     *
      * @param config 参数配置信息
      * @return 参数配置集合
      */
@@ -36,7 +36,7 @@ public interface SysConfigMapper
 
     /**
      * 根据键名查询参数配置信息
-     * 
+     *
      * @param configKey 参数键名
      * @return 参数配置信息
      */
@@ -44,7 +44,7 @@ public interface SysConfigMapper
 
     /**
      * 新增参数配置
-     * 
+     *
      * @param config 参数配置信息
      * @return 结果
      */
@@ -52,7 +52,7 @@ public interface SysConfigMapper
 
     /**
      * 修改参数配置
-     * 
+     *
      * @param config 参数配置信息
      * @return 结果
      */
@@ -60,17 +60,17 @@ public interface SysConfigMapper
 
     /**
      * 删除参数配置
-     * 
+     *
      * @param configId 参数ID
      * @return 结果
      */
-    public int deleteConfigById(Long configId);
+    public int deleteConfigById(Integer configId);
 
     /**
      * 批量删除参数信息
-     * 
+     *
      * @param configIds 需要删除的参数ID
      * @return 结果
      */
-    public int deleteConfigByIds(Long[] configIds);
+    public int deleteConfigByIds(Integer[] configIds);
 }
