+++ ruoyi-modules/ruoyi-system/src/main/java/com/ruoyi/system/mapper/SysDictTypeMapper.java
@@ -1,18 +1,18 @@
 package com.ruoyi.system.mapper;
 
 import java.util.List;
+
 import com.ruoyi.system.api.domain.SysDictType;
 
 /**
  * 字典表 数据层
- * 
+ *
  * @author ruoyi
  */
-public interface SysDictTypeMapper
-{
+public interface SysDictTypeMapper {
     /**
      * 根据条件分页查询字典类型
-     * 
+     *
      * @param dictType 字典类型信息
      * @return 字典类型集合信息
      */
@@ -20,22 +20,22 @@ public interface SysDictTypeMapper
 
     /**
      * 根据所有字典类型
-     * 
+     *
      * @return 字典类型集合信息
      */
     public List<SysDictType> selectDictTypeAll();
 
     /**
      * 根据字典类型ID查询信息
-     * 
+     *
      * @param dictId 字典类型ID
      * @return 字典类型
      */
-    public SysDictType selectDictTypeById(Long dictId);
+    public SysDictType selectDictTypeById(Integer dictId);
 
     /**
      * 根据字典类型查询信息
-     * 
+     *
      * @param dictType 字典类型
      * @return 字典类型
      */
@@ -43,23 +43,23 @@ public interface SysDictTypeMapper
 
     /**
      * 通过字典ID删除字典信息
-     * 
+     *
      * @param dictId 字典ID
      * @return 结果
      */
-    public int deleteDictTypeById(Long dictId);
+    public int deleteDictTypeById(Integer dictId);
 
     /**
      * 批量删除字典类型信息
-     * 
+     *
      * @param dictIds 需要删除的字典ID
      * @return 结果
      */
-    public int deleteDictTypeByIds(Long[] dictIds);
+    public int deleteDictTypeByIds(Integer[] dictIds);
 
     /**
      * 新增字典类型信息
-     * 
+     *
      * @param dictType 字典类型信息
      * @return 结果
      */
@@ -67,7 +67,7 @@ public interface SysDictTypeMapper
 
     /**
      * 修改字典类型信息
-     * 
+     *
      * @param dictType 字典类型信息
      * @return 结果
      */
@@ -75,7 +75,7 @@ public interface SysDictTypeMapper
 
     /**
      * 校验字典类型称是否唯一
-     * 
+     *
      * @param dictType 字典类型
      * @return 结果
      */
