+++ ruoyi-modules/ruoyi-system/src/main/java/com/ruoyi/system/service/ISysOperLogService.java
@@ -1,18 +1,18 @@
 package com.ruoyi.system.service;
 
 import java.util.List;
+
 import com.ruoyi.system.api.domain.SysOperLog;
 
 /**
  * 操作日志 服务层
- * 
+ *
  * @author ruoyi
  */
-public interface ISysOperLogService
-{
+public interface ISysOperLogService {
     /**
      * 新增操作日志
-     * 
+     *
      * @param operLog 操作日志对象
      * @return 结果
      */
@@ -20,7 +20,7 @@ public interface ISysOperLogService
 
     /**
      * 查询系统操作日志集合
-     * 
+     *
      * @param operLog 操作日志对象
      * @return 操作日志集合
      */
@@ -28,19 +28,19 @@ public interface ISysOperLogService
 
     /**
      * 批量删除系统操作日志
-     * 
+     *
      * @param operIds 需要删除的操作日志ID
      * @return 结果
      */
-    public int deleteOperLogByIds(Long[] operIds);
+    public int deleteOperLogByIds(Integer[] operIds);
 
     /**
      * 查询操作日志详细
-     * 
+     *
      * @param operId 操作ID
      * @return 操作日志对象
      */
-    public SysOperLog selectOperLogById(Long operId);
+    public SysOperLog selectOperLogById(Integer operId);
 
     /**
      * 清空操作日志
